Add Flickr API

+++ app/src/main/java/com/kristinaefros/challenge/data/storage/entity/PlaceEntity.kt
@@ -12,8 +12,8 @@ data class PlaceEntity(
     @PrimaryKey(autoGenerate = true) val id: Int,
     val latitude: Double,
     val longitude: Double,
-    val radius: Double,
+    val radius: Float,
     val photoUrl: String?,
 ) {
-    constructor(latitude: Double, longitude: Double, radius: Double) : this(0, latitude, longitude, radius, null)
+    constructor(latitude: Double, longitude: Double, radius: Float) : this(0, latitude, longitude, radius, null)
 }
